refactor: re-write gen_gql_schema.rs #1510
