+++ tailcall-typedefs/src/sservice-document.rs
@@ -1,0 +1,23 @@
+use anyhow::Result;
+use async_graphql::{ServiceDocument, ObjectType};
+use schemars::schema::{Schema, SchemaObject};
+use std::collections::BTreeMap;
+
+pub fn generate_service_document() -> Result<ServiceDocument> {
+    let schema = schemars::schema_for!(Config);
+    let defs: BTreeMap<String, Schema> = schema.definitions;
+    
+    let mut service_document = ServiceDocument::new();
+    
+    for (name, schema) in defs.iter() {
+        let object_type = convert_schema_to_object_type(schema.clone().into_object())?;
+        service_document.add_type(name, object_type);
+    }
+    
+    Ok(service_document)
+}
+
+fn convert_schema_to_object_type(schema: SchemaObject) -> Result<ObjectType> {
+    // Conversion logic here
+    Ok(ObjectType::new("example_type"))
+}
