+++ tailcall-typedefs/src/utils.rs
@@ -1,0 +1,13 @@
+pub fn first_char_to_upper(name: &mut String) {
+    if let Some(first_char) = name.chars().next() {
+        // Remove the first character and make it uppercase
+        let first_char_upper = first_char.to_uppercase().to_string();
+
+        // Remove the first character from the original string
+        let mut chars = name.chars();
+        chars.next();
+
+        // Replace the original string with the new one
+        *name = first_char_upper + chars.as_str();
+    }
+}
